Pull export logic out of component file
